+++ ProduktEingabe.cs
@@ -7,7 +7,7 @@ namespace WpfApp
         public string Produkt { get; set; }
         public string Beschreibung { get; set; }
         public decimal Preis { get; set; }
-        public int Anzahl { get; set; }
+        public decimal Anzahl { get; set; }
         public decimal Gesamtpreis { get; set; }
 
     }
